feat: added column names, tasks controller - addTaskType, turn off email

+++ src/main/java/com/group/contestback/repositories/AppUserRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface AppUserRepo extends JpaRepository<AppUser, Long> {
+public interface AppUserRepo extends JpaRepository<AppUser, Integer> {
     //AppUser findByUsername(String username);
     Optional<AppUser> findById(Integer id);
     AppUser findByLogin(String login);
